# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "API boilerplate with Go + Fiber.",
        "title": "Boilerplate Go Fiber API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/payment/balance": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Check Balance",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/internal_transport_http_payment.BalanceResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                        }
                    }
                }
            }
        },
        "/payment/invoice": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Create Invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Idempotency key",
                        "name": "Idempotency-Key",
                        "in": "header"
                    },
                    {
                        "description": "Create invoice payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_transport_http_payment.CreateInvoiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/invoice.Invoice"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/invoice.Invoice"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                        }
                    }
                }
            }
        },
        "/payment/invoice/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Get Invoice By Id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/invoice.Invoice"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                        }
                    }
                }
            }
        },
        "/payment/invoice/{id}/expire": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Expire Invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/invoice.Invoice"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                        }
                    }
                }
            }
        },
        "/payment/webhook/xendit/invoice": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Payment"
                ],
                "summary": "Xendit invoice webhook",
                "parameters": [
                    {
                        "description": "Invoice callback payload",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/invoice.InvoiceCallback"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_mzulfanw_boilerplate-go-fiber_internal_transport_http_response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "internal_transport_http_payment.BalanceResponse": {
            "type": "object",
            "properties": {
                "balance": {
                    "type": "number"
                }
            }
        },
        "internal_transport_http_payment.CreateInvoiceRequest": {
            "type": "object",
            "required": [
                "amount",
                "external_id"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "callback_virtual_account_id": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "failure_redirect_url": {
                    "type": "string"
                },
                "fees": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_transport_http_payment.InvoiceFeeRequest"
                    }
                },
                "invoice_duration": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_transport_http_payment.InvoiceItemRequest"
                    }
                },
                "locale": {
                    "type": "string"
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "mid_label": {
                    "type": "string"
                },
                "payer_email": {
                    "type": "string"
                },
                "payment_methods": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "reminder_time": {
                    "type": "integer"
                },
                "reminder_time_unit": {
                    "type": "string"
                },
                "should_authenticate_credit_card": {
                    "type": "boolean"
                },
                "should_send_email": {
                    "type": "boolean"
                },
                "success_redirect_url": {
                    "type": "string"
                }
            }
        },
        "internal_transport_http_payment.InvoiceFeeRequest": {
            "type": "object",
            "required": [
                "type",
                "value"
            ],
            "properties": {
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "number",
                    "minimum": 0
                }
            }
        },
        "internal_transport_http_payment.InvoiceItemRequest": {
            "type": "object",
            "required": [
                "name",
                "price",
                "quantity"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number",
                    "minimum": 0
                },
                "quantity": {
                    "type": "number",
                    "minimum": 0
                },
                "reference_id": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "invoice.Bank": {
            "type": "object",
            "properties": {
                "account_holder_name": {
                    "description": "The name of the account holder.",
                    "type": "string"
                },
                "alternative_displays": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoice.AlternativeDisplayItem"
                    }
                },
                "bank_account_number": {
                    "description": "The bank account number.",
                    "type": "string"
                },
                "bank_branch": {
                    "description": "The branch of the bank.",
                    "type": "string"
                },
                "bank_code": {
                    "$ref": "#/definitions/invoice.BankCode"
                },
                "collection_type": {
                    "description": "The collection type for the bank details.",
                    "type": "string"
                },
                "transfer_amount": {
                    "description": "The transfer amount.",
                    "type": "number"
                }
            }
        },
...
        "invoice.ChannelProperties": {
            "type": "object",
            "properties": {
                "cards": {
                    "$ref": "#/definitions/invoice.ChannelPropertiesCards"
                }
            }
        },
...
        "invoice.DirectDebit": {
            "type": "object",
            "properties": {
                "direct_debit_type": {
                    "$ref": "#/definitions/invoice.DirectDebitType"
                }
            }
        },
...
        "invoice.Ewallet": {
            "type": "object",
            "properties": {
                "ewallet_type": {
                    "$ref": "#/definitions/invoice.EwalletType"
                }
            }
        },
...
        "invoice.Invoice": {
            "type": "object",
            "properties": {
                "amount": {
                    "description": "The total amount of the invoice.",
                    "type": "number"
                },
                "available_banks": {
                    "description": "An array of available banks for payment.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoice.Bank"
                    }
                },
                "available_direct_debits": {
                    "description": "An array of available direct debit options for payment.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoice.DirectDebit"
                    }
                },
                "available_ewallets": {
                    "description": "An array of available e-wallets for payment.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoice.Ewallet"
                    }
                },
                "available_paylaters": {
                    "description": "An array of available pay-later options for payment.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoice.Paylater"
                    }
                },
                "available_qr_codes": {
                    "description": "An array of available QR codes for payment.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoice.QrCode"
                    }
                },
                "available_retail_outlets": {
                    "description": "An array of available retail outlets for payment.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoice.RetailOutlet"
                    }
                },
                "channel_properties": {
                    "$ref": "#/definitions/invoice.ChannelProperties"
                },
                "created": {
                    "description": "Representing a date and time in ISO 8601 format.",
                    "type": "string"
                },
                "currency": {
                    "$ref": "#/definitions/invoice.InvoiceCurrency"
                },
                "customer": {
                    "$ref": "#/definitions/invoice.CustomerObject"
                },
                "customer_notification_preference": {
                    "$ref": "#/definitions/invoice.NotificationPreference"
                },
                "description": {
                    "description": "A description of the invoice.",
                    "type": "string"
                },
                "expiry_date": {
                    "description": "Representing a date and time in ISO 8601 format.",
                    "type": "string"
                },
                "external_id": {
                    "description": "The external identifier for the invoice.",
                    "type": "string"
                },
                "failure_redirect_url": {
                    "description": "The URL to redirect to on payment failure.",
                    "type": "string"
                },
                "fees": {
                    "description": "An array of fees associated with the invoice.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoice.InvoiceFee"
                    }
                },
                "fixed_va": {
                    "description": "Indicates whether the virtual account is fixed.",
                    "type": "boolean"
                },
                "id": {
                    "description": "The unique identifier for the invoice.",
                    "type": "string"
                },
                "invoice_url": {
                    "description": "The URL to view the invoice.",
                    "type": "string"
                },
                "items": {
                    "description": "An array of items included in the invoice.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/invoice.InvoiceItem"
                    }
                },
                "locale": {
                    "description": "The locale or language used for the invoice.",
                    "type": "string"
                },
                "merchant_name": {
                    "description": "The name of the merchant.",
                    "type": "string"
                },
                "merchant_profile_picture_url": {
                    "description": "The URL of the merchant's profile picture.",
                    "type": "string"
                },
                "metadata": {
                    "description": "A free-format JSON for additional information that you may use. Object can be up to 50 keys, with key names up to 40 characters long and values up to 500 characters long.",
                    "type": "object",
                    "additionalProperties": true
                },
                "payer_email": {
                    "description": "The email address of the payer.",
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/invoice.InvoicePaymentMethod"
                },
                "reminder_date": {
                    "description": "Representing a date and time in ISO 8601 format.",
                    "type": "string"
                },
                "should_authenticate_credit_card": {
                    "description": "Indicates whether credit card authentication is required.",
                    "type": "boolean"
                },
                "should_exclude_credit_card": {
                    "description": "Indicates whether credit card payments should be excluded.",
                    "type": "boolean"
                },
                "should_send_email": {
                    "description": "Indicates whether email notifications should be sent.",
                    "type": "boolean"
                },
                "status": {
                    "$ref": "#/definitions/invoice.InvoiceStatus"
                },
                "success_redirect_url": {
                    "description": "The URL to redirect to on successful payment.",
                    "type": "string"
                },
                "updated": {
                    "description": "Representing a date and time in ISO 8601 format.",
                    "type": "string"
                },
                "user_id": {
                    "description": "The user ID associated with the invoice.",
                    "type": "string"
                }
            }
        },
...
        "invoice.Paylater": {
            "type": "object",
            "properties": {
                "paylater_type": {
                    "$ref": "#/definitions/invoice.PaylaterType"
                },
                "should_exclude": {
                    "description": "Indicates whether this paylater option should be excluded.",
                    "type": "boolean"
                }
            }
        },
...
        "invoice.QrCode": {
            "type": "object",
            "properties": {
                "qr_code_type": {
                    "$ref": "#/definitions/invoice.QrCodeType"
                }
            }
        },
...
        "invoice.RetailOutlet": {
            "type": "object",
            "properties": {
                "merchant_name": {
                    "description": "The name of the merchant.",
                    "type": "string"
                },
                "payment_code": {
                    "description": "The payment code.",
                    "type": "string"
                },
                "retail_outlet_name": {
                    "$ref": "#/definitions/invoice.RetailOutletName"
                },
                "transfer_amount": {
                    "description": "The transfer amount.",
                    "type": "number"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}